Add PythonEditor subclass
